# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI4BMI API Groupe 3",
        "description": "API REST pour le système BMI — Gestion équipements et E-commerce BMI",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/v1/cart": {
            "get": {
                "tags": [
                    "Panier"
                ],
                "summary": "Voir mon panier",
                "description": "Retourne la liste des articles dans le panier de l'utilisateur authentifié, incluant les prix unitaires et le total général.",
                "operationId": "2721c2a7f5993ff11ef640d1ca9a2c7c",
                "responses": {
                    "200": {
                        "description": "Contenu du panier récupéré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "cart_id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "quantity": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "product": {
                                                                "properties": {
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "Pneu Michelin"
                                                                    },
                                                                    "price": {
                                                                        "type": "number",
                                                                        "example": 45000
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "string",
                                                    "example": "90,000.00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Panier"
                ],
                "summary": "Vider tout le panier",
                "description": "Supprime l'intégralité des articles du panier en une seule action.",
                "operationId": "6fb158682b5cd7446fa485ffa0de60c7",
                "responses": {
                    "200": {
                        "description": "Panier réinitialisé avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/cart/items": {
            "post": {
                "tags": [
                    "Panier"
                ],
                "summary": "Ajouter un article",
                "description": "Ajoute un produit au panier. Bloque l'ajout si le produit est inactif ou si le stock est insuffisant.",
                "operationId": "c16629a212180f36fb1c85298ec00d69",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "description": "ID du produit à ajouter",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "description": "Nombre d'unités souhaitées",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produit ajouté avec succès"
                    },
                    "400": {
                        "description": "Stock insuffisant ou produit indisponible"
                    },
                    "422": {
                        "description": "Erreur de validation des paramètres"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/cart/items/{cartItemId}": {
            "put": {
                "tags": [
                    "Panier"
                ],
                "summary": "Modifier la quantité",
                "description": "Met à jour la quantité d'un article spécifique via son ID de ligne de panier.",
                "operationId": "ffc5bb205658709f6b392fc7271d6206",
                "parameters": [
                    {
                        "name": "cartItemId",
                        "in": "path",
                        "description": "ID de la ligne du panier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quantité mise à jour"
                    },
                    "404": {
                        "description": "L'article n'est pas dans votre panier"
                    },
                    "400": {
                        "description": "Le stock est insuffisant pour cette nouvelle quantité"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Panier"
                ],
                "summary": "Supprimer un article",
                "description": "Retire définitivement une ligne de produit du panier.",
                "operationId": "f416c58d581621bc79814764c19c76f5",
                "parameters": [
                    {
                        "name": "cartItemId",
                        "in": "path",
                        "description": "ID de la ligne du panier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article retiré avec succès"
                    },
                    "404": {
                        "description": "Article introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
}